processor object
