+++ sensor/common/sensor/central_receiver.go
@@ -4,7 +4,6 @@ import (
 	"github.com/stackrox/rox/generated/internalapi/central"
 	"github.com/stackrox/rox/pkg/concurrency"
 	"github.com/stackrox/rox/pkg/sync"
-	"github.com/stackrox/rox/sensor/common"
 )
 
 // CentralReceiver handles receiving data from central.
@@ -15,7 +14,7 @@ type CentralReceiver interface {
 }
 
 // NewCentralReceiver returns a new instance of a Receiver.
-func NewCentralReceiver(finished *sync.WaitGroup, receivers ...common.SensorComponent) CentralReceiver {
+func NewCentralReceiver(finished *sync.WaitGroup, receivers Processor) CentralReceiver {
 	return &centralReceiverImpl{
 		stopper:   concurrency.NewStopper(),
 		receivers: receivers,
