+++ sensor/kubernetes/sensor/sensor.go
@@ -179,6 +179,8 @@ func CreateSensor(cfg *CreateOptions) (*sensor.Sensor, error) {
 			certrefresh.NewSecuredClusterTLSIssuer(cfg.introspectionK8sClient.Kubernetes(), sensorNamespace, podName))
 	}
 
+	processor := sensor.NewProcessor(components...)
+
 	s := sensor.NewSensor(
 		configHandler,
 		policyDetector,
@@ -186,6 +188,7 @@ func CreateSensor(cfg *CreateOptions) (*sensor.Sensor, error) {
 		cfg.centralConnFactory,
 		pubSub,
 		cfg.certLoader,
+		processor,
 		components...,
 	)
